Move as_pyam from .compat.pyam to .compat.pyam.comp
